requirements/common.txt
alembic==0.8.8
bleach==1.4.3
blinker==1.4
click==6.6
dominate==2.2.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
List==1.3.0
Mako==1.0.4
Markdown==2.6.7
MarkupSafe==0.23
Pillow==3.4.1
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.15
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
